requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15,<4.0
pymongo>=4.8,<5.0
httpx>=0.27,<1.0
pydantic>=2.8,<3.0
gunicorn>=22.0,<23.0
pytest>=8.3,<9.0
pytest-django>=4.9,<5.0
mongomock>=4.1,<5.0
